// frontend/src/app/core/models/student.model.ts
export interface Student {
  studentRegno: string;
  studentPhoto?: string;
  password?: string;
  studentName?: string;
  regno?: string;
  name?: string;
  photo?: string;
  email?: string;
  pincode?: string;
  session?: string;
  department?: string;
  level?: string;
  semester?: string;
  cgpa?: number;
  creationdate?: string;
  updationDate?: string;
}